+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [1.7.9] - 2025-04-16
+
 ### Changed
 
 - Support for .NET Aspire 9.2
@@ -377,7 +379,8 @@
 
 - Support for running and debugging of Aspire projects
 
-[Unreleased]: https://github.com/JetBrains/aspire-plugin/compare/v1.7.8...HEAD
+[Unreleased]: https://github.com/JetBrains/aspire-plugin/compare/v1.7.9...HEAD
+[1.7.9]: https://github.com/JetBrains/aspire-plugin/compare/v1.7.8...v1.7.9
 [1.7.8]: https://github.com/JetBrains/aspire-plugin/compare/v1.7.7...v1.7.8
 [1.7.7]: https://github.com/JetBrains/aspire-plugin/compare/v1.7.6...v1.7.7
 [1.7.6]: https://github.com/JetBrains/aspire-plugin/compare/v1.7.5...v1.7.6